process logic for app

+++ lib/model/wind.dart
@@ -5,7 +5,7 @@ class Wind {
   Wind({this.speed, this.deg});
 
   factory Wind.fromJson(Map<String, dynamic> json) =>
-      Wind(speed: json["speed"], deg: json["deg"].toDouble());
+      Wind(speed: json["speed"].toDouble(), deg: json["deg"].toDouble());
 
   Map<String, dynamic> toJson() => {"speed": speed, "deg": deg};
 }
